refactor(plugins): move cronPlugin to src/plugins/cron/ (CP-1)
- src/lib/cron-plugin.ts -> src/plugins/cron/cron-plugin.ts
- 1 import site updated (src/index.ts). Internal imports from
  @/lib/logger and @/modules/... unchanged.

Tests: 137 pass across payments/jobs and occurrences/vacations
(cronPlugin consumers) — no runtime cycle after static-import promotion
from CP-30 and relocation.

CP-32 (declarative array for the 7 inline cron jobs) stays out of
scope — tracked as separate CP.
